requirements.txt
mysql-connector==2.2.9
numpy==1.17.2
Pillow==6.2.0
PyMySQL==0.9.3
pyodbc==4.0.27
